Fixup tests on mac including asset symlinks support (#314)
This updates the webpack-asset-relocator-loader to support the tfjs-node build that was failing on MacOS.

The failure was due to shared libraries with .so extensions not begin emitted on mac, and also due to not having support for symlinked assets, as the pathing of the shared library relies on symlinks. Both of these are fixed in the latest webpack-asset-relocator-loader release at https://github.com/zeit/webpack-asset-relocator-loader/commit/23b0dbffce22fbe19f51c292a73f40ba82ab8386.