fix: incorrect usage of Hash.digest
Hash.digest expect an encoding value as parameter(latin1 | hex | base64).
The absolute path of the file was passed to Hash.digest which is incorrect. To generate a hash
based on the pathname we have to call Hash.update first.
The desired encoding format is then passed to Hash.digest. This remove
the need to call toString method.